Tests/LibWeb/Text/expected/HTML/image-load-after-adoption-from-removed-iframe.txt

image loaded after adoption
width=120 height=120 naturalWidth=120 naturalHeight=120 complete=true
